Rename `folding::error` to `error_term`
